agents/screeps-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "screeps-agent"
version = "0.1.0"
description = "AI Agent for Screeps game based on kimi-cli"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kimi-cli>=0.3.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "aiohttp>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.scripts]
screeps-agent = "screeps_agent.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/screeps_agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
